moving serialize_variable_values into utilities folder
