+++ gql/utilities/__init__.py
@@ -1,4 +1,5 @@
 from .parse_result import parse_result
+from .serialize_variable_values import serialize_value, serialize_variable_values
 from .update_schema_enum import update_schema_enum
 from .update_schema_scalars import update_schema_scalar, update_schema_scalars
 
@@ -7,4 +8,6 @@ __all__ = [
     "update_schema_scalar",
     "update_schema_enum",
     "parse_result",
+    "serialize_variable_values",
+    "serialize_value",
 ]
