+++ tests/custom_scalars/test_money.py
@@ -21,8 +21,7 @@ from graphql.utilities import value_from_ast_untyped
 
 from gql import Client, gql
 from gql.transport.exceptions import TransportQueryError
-from gql.utilities import update_schema_scalar, update_schema_scalars
-from gql.variable_values import serialize_value
+from gql.utilities import serialize_value, update_schema_scalar, update_schema_scalars
 
 from ..conftest import MS
